种地模块完成。

+++ haoyou-spring-cloud-alibaba-04-commons-mapper/src/main/java/com/haoyou/spring/cloud/alibaba/commons/entity/Currency.java
@@ -20,12 +20,12 @@ public class Currency implements Serializable {
     /**
      * 金币
      */
-    private Integer coin;
+    private Long coin;
 
     /**
      * 钻石
      */
-    private Integer diamond;
+    private Long diamond;
 
     /**
      * 体力
