+++ haoyou-spring-cloud-alibaba-cultivate/src/main/java/com/haoyou/spring/cloud/alibaba/cultivate/currency/use/handle/CultureLimitHandle.java
@@ -62,8 +62,8 @@ public class CultureLimitHandle extends CurrencyUseHandle {
         }
 
         //所需道具
-        int propCount = cultureLimit / 10 * 1 * cultureLimit + (cultureLimit % 10)+10;
-        int nprop = cultureMedium.getCount() - propCount;
+        long propCount = cultureLimit / 10 * 1 * cultureLimit + (cultureLimit % 10)+10;
+        long nprop = cultureMedium.getCount() - propCount;
         if(nprop <0){
             return PROP_LESS;
         }else{
@@ -71,8 +71,8 @@ public class CultureLimitHandle extends CurrencyUseHandle {
         }
 
         //所需钻石
-        int diamondCount = cultureLimit / 10 * 10 * cultureLimit + (cultureLimit % 10)*10+100;
-        int nDiamond = user.getCurrency().getDiamond() - diamondCount;
+        long diamondCount = cultureLimit / 10 * 10 * cultureLimit + (cultureLimit % 10)*10+100;
+        long nDiamond = user.getCurrency().getDiamond() - diamondCount;
         if (nDiamond < 0) {
             return DIAMOND_LESS;
         } else {
